requirements/libraries.txt
# Mandatory (packaged) libraries
aenum==2.1.2
appdirs==1.4.3
attrs==18.2.0
babelfish==0.5.5
backports.functools_lru_cache=1.5 ; python_version < '3.0'
beautifulsoup4==4.7.1 # different package for py2 and py3
boto==2.49.0
certifi==2018.11.29
chardet==3.0.4
cheroot==6.5.4
cherrypy==17.4.1 # rq.filter: <18.0.0
click==7.0
configparser==3.5.0
contextlib2=0.5.5
decorator==4.3.0
diskcache==3.1.1
dogpile.cache==0.7.1
enzyme==0.4.1
futures==3.2.0 ; python_version < '3.0'
gitdb2==2.0.5
gitpython==2.1.11
gntp==1.0.3
guessit==2.1.4 # rq.filter: <3.0.0
html5lib==1.0.1
idna==2.8
imdbpie==5.6.3
jaraco.functools==2.0
jsonpickle==1.0
langdetect==1.0.7
lxml==4.3.0
mako==1.0.7
markupsafe==1.1.0
more-itertools==5.0.0
oauthlib==3.0.0
portend==2.3
pushbullet.py==0.11.0
pysrt==1.1.1
python-dateutil==2.7.5
python-magic==0.4.15
python-twitter==3.5
pytz==2018.9
rarfile==3.0
rebulk==1.0.0
requests==2.21.0
requests-oauthlib==1.1.0
schema==0.6.8
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
soupsieve==1.6.2
stevedore==1.30.0
git+https://github.com/Diaoul/subliminal.git@879c0c8d18618e32ad466bb63bd939edb268bc60#egg=subliminal
tempora==1.14
trans==2.1.0
git+https://github.com/h3llrais3r/tvdb_api_v2.git@7f02ecb9620017aed75973fd14d0197825c35140#egg=tvdb_api_v2
unidecode==1.0.23
urllib3==1.24.1
webencodings==0.5.1
websocket-client==0.54.0
ws4py==0.5.1
zc.lockfile==1.4
